requirements.txt
easyocr==1.7.1
numpy==1.26.4
pillow==10.3.0
streamlit==1.35.0
watchdog==4.0.1
